+++ CHANGELOG.md
@@ -14,7 +14,6 @@ For information on changes in released versions of Teku, see the [releases page]
 ### Breaking Changes
 
 ### Additions and Improvements
- - Optimized epoch transitions
+ - Log a warning if node is syncing while performing sync committee duties instead of a verbose error
 
 ### Bug Fixes
- - Fixed the beacon-rest-api opening a listen port on too many interfaces